- Implement Zustand cart slice with add and remove functionality
This commit introduces a Zustand slice named cartSlice that handles the cart-related state management. The slice includes functions to add and remove items from the cart, as well as the necessary state variables to track cart items and total quantities. The addToCart function checks whether an item already exists in the cart and either increments its quantity or adds it to the cart. Similarly, the removeFromCart function handles reducing an item's quantity or removing it from the cart entirely based on user interaction.

- Implement AddToCartBtn Component with Zustand Integration

This commit introduces the AddToCartBtn component, which allows users to add items to their cart. The component uses Zustand for state management. It dynamically adjusts its appearance and behavior based on whether the item is already in the cart or not. If the item is in the cart, the component displays buttons to increase or decrease the item quantity. Otherwise, it provides an "ADD TO CART" button.

The handleAddToCart function is triggered when the user clicks the "ADD TO CART" button. It calls the addToCart action provided by the Zustand store and passes the necessary item details, including _id, title, brand, category, description, image, isNew, oldPrice, price, and an initial quantity of 1.

- Implement the total number of cart and favorites dynamically in the Header component